requirements.txt
asgiref==3.2.3
<<<<<<< HEAD
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-wkhtmltopdf==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
zipp==3.0.0
=======
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.1
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
>>>>>>> 03e218cf9638771148b7fc638f13fc66fa822b6e
